validations and exception handler

+++ Services/Impl/AuthorService .cs
@@ -2,6 +2,7 @@
 
 using books_api.Data.Repositories;
 using Services;
+using books_api.Exceptions;
 
 namespace books_api.Services.Impl
 {
@@ -12,7 +13,6 @@ namespace books_api.Services.Impl
         public AuthorService(IAuthorRepository authorRepository)
         {
             _authorRepository = authorRepository;
-
         }
 
         public async Task<IEnumerable<Author>> GetAllAsync()
@@ -42,7 +42,10 @@ namespace books_api.Services.Impl
 
         public async Task DeleteAsync(Guid id)
         {
+            if (await _authorRepository.AuthorInUseAsync(id))
+                throw new EntityInUseException($"Author with id: '{id}' is used in some book.");
             Author existentAuthor = await FindAsync(id);
+
             await _authorRepository.DeleteAsync(existentAuthor);
         }
     }
